setools: Do not use old setup.py as setup file. Remove old setup.py.
There is a comment about new setools not being modular but no
explanation why we should care about it and why we should use setup.py
from older version. Thus, removed setting of the old setup.py.

Removed the file as there is no use for it.

Also cleaned up parts of the spec file.