[setools] Update to 4.4.4. JB#62619
python3-networkx was made optional so we shouldn't be required to
apply the patch anymore. Furthermore the earlier version wasn't properly
filtering out the dependency so the package wasn't installable.

Tried 4.5.1 too but that was now having some cython problem. Went the
easier route and did an intermediate update for now, it's still
relatively new.